+++ app/page.tsx
@@ -1,31 +1,16 @@
 "use client";
 
 import { experimental_useObject as useObject } from "ai/react";
-import { CardSchema, CardSchemaType } from "../lib/cardSchema";
+import { CardSchema, CardSchemaType } from "@/lib/cardSchema";
 import { useCallback, useEffect, useRef, useState } from "react";
-import {
-  History,
-  parseHistory,
-  serializeHistory,
-  toSerializableHistory,
-} from "../lib/history";
-import { Button } from "@/components/ui/button";
 import { SelectCategory } from "@/components/ui/select-category";
 import { Switch } from "@/components/ui/switch";
 import { Label } from "@/components/ui/label";
-import { DraggableCard } from "@/components/ui/draggable-card";
-
-function createRequestBody(categoryName: string, history: History) {
-  const requestBody = {
-    categoryName,
-    history: toSerializableHistory(history),
-  };
-
-  return requestBody;
-}
+import { History, parseHistory, serializeHistory } from "@/lib/history";
+import { createRequestBody } from "@/lib/utils";
+import { CardDeck } from "@/components/ui/card-deck";
 
 export default function Home() {
-  const [conceptsHistory, setConceptsHistory] = useState<History>(new Map());
   const [categoryName, setCategoryName] = useState("");
   const [isTesting, setIsTesting] = useState(false);
   const { object, submit, isLoading, error } = useObject({
@@ -35,30 +20,14 @@ export default function Home() {
       names: [],
     },
   });
+  const [conceptsHistory, setConceptsHistory] = useState<History>(new Map());
   const [isHistoryLoaded, setIsHistoryLoaded] = useState(false);
+
   const isInitialFetch = useRef(true);
 
   const fetchData = useCallback(() => {
     submit(createRequestBody(categoryName, conceptsHistory));
   }, [categoryName, conceptsHistory, submit]);
-
-  useEffect(() => {
-    const savedHistory = localStorage.getItem("conceptsHistory");
-    if (savedHistory) {
-      setConceptsHistory(parseHistory(savedHistory));
-    } else {
-      setConceptsHistory(new Map());
-    }
-
-    setIsHistoryLoaded(true);
-  }, []);
-
-  useEffect(() => {
-    if (isInitialFetch.current && isHistoryLoaded) {
-      fetchData();
-      isInitialFetch.current = false;
-    }
-  }, [fetchData, isHistoryLoaded]);
 
   useEffect(() => {
     if ((object as CardSchemaType).names.length > 0) {
@@ -84,15 +53,35 @@ export default function Home() {
     }
   }, [object]);
 
+  useEffect(() => {
+    if (isInitialFetch.current && isHistoryLoaded) {
+      fetchData();
+      isInitialFetch.current = false;
+    }
+  }, [fetchData, isHistoryLoaded]);
+
+  useEffect(() => {
+    const savedHistory = localStorage.getItem("conceptsHistory");
+    if (savedHistory) {
+      setConceptsHistory(parseHistory(savedHistory));
+    } else {
+      setConceptsHistory(new Map());
+    }
+
+    setIsHistoryLoaded(true);
+  }, []);
+
   return (
     <main className="min-w-0 w-full max-w-[46ch] py-10 p-6 flex flex-col gap-4 items-center">
-      {/* <CardDeck /> */}
-      <DraggableCard
-        object={{ names: object?.names ?? [] } as CardSchemaType}
-        isLoading={isLoading}
-        isTesting={isTesting}
-      />
-      <div className="w-full space-y-4">
+      {!error && (
+        <CardDeck
+          object={{ names: object?.names ?? [] } as CardSchemaType}
+          isTesting={isTesting}
+          isLoading={isLoading}
+        />
+      )}
+
+      <div className="w-full">
         {isTesting && (
           <SelectCategory
             isLoading={isLoading}
@@ -100,13 +89,6 @@ export default function Home() {
             onValueChange={setCategoryName}
           />
         )}
-        <Button
-          className="w-full"
-          disabled={isLoading}
-          onClick={() => fetchData()}
-        >
-          New Card
-        </Button>
       </div>
 
       {isTesting && (
@@ -121,6 +103,7 @@ export default function Home() {
           </a>
         </p>
       )}
+
       <div className="flex items-center gap-2">
         <Switch
           id="testing-mode-switch"
